# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Elite Care Medical API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  inStock: boolean;
  isFeatured: boolean;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  specifications?: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  slug: string;
  description?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  imageUrl?: string;
  images?: string[];
  inStock?: boolean;
  isFeatured?: boolean;
  brand?: string;
  model?: string;
  specifications?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId?: number;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  inStock?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  specifications?: string | null;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  items: OrderItem[];
  subtotal: number;
  deliveryFee?: number;
  total: number;
  status: string;
  paymentStatus: string;
  paymentMethod: string;
  /** @nullable */
  paystackReference?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerEmail: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  deliveryAddress: string;
  city?: string;
  state?: string;
  items: OrderItemInput[];
  paymentMethod: string;
  notes?: string;
}

export interface OrderStatusUpdate {
  status: string;
  paymentStatus?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
}

export interface PaymentInit {
  orderId: number;
  email: string;
  amount: number;
  callbackUrl?: string;
}

export interface PaymentInitResponse {
  authorizationUrl: string;
  reference: string;
  accessCode: string;
}

export interface PaymentVerifyResponse {
  status: string;
  message: string;
  orderId: number;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminAuthToken {
  success: boolean;
  token: string;
}

export interface CategoryStat {
  categoryName: string;
  productCount: number;
  orderCount: number;
}

export interface AdminStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  categoryBreakdown: CategoryStat[];
  recentOrders: Order[];
}

export interface WhatsAppStatus {
  configured: boolean;
  phoneNumberConfigured: boolean;
  verifyTokenConfigured: boolean;
  appSecretConfigured: boolean;
}

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
featured?: boolean | null;
};

export type ListOrdersParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
page?: number | null;
};

export type VerifyWhatsAppWebhookParams = {
'hub.mode': string;
'hub.verify_token': string;
'hub.challenge': string;
};

export type ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValueMessagesItemText = {
  body?: string;
};

export type ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValueMessagesItem = {
  from?: string;
  id?: string;
  type?: string;
  text?: ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValueMessagesItemText;
};

export type ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValue = {
  messages?: ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValueMessagesItem[];
};

export type ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItem = {
  value?: ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItemValue;
};

export type ReceiveWhatsAppWebhookPayloadBodyEntryItem = {
  changes?: ReceiveWhatsAppWebhookPayloadBodyEntryItemChangesItem[];
};

export type ReceiveWhatsAppWebhookPayloadBody = {
  object?: string;
  entry?: ReceiveWhatsAppWebhookPayloadBodyEntryItem[];
};