# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elite Care Medical API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    Order:
      type: object
      required: [id, customerName, customerEmail, customerPhone, deliveryAddress, items, subtotal, total, status, paymentStatus, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        status:
          type: string
        paymentStatus:
          type: string
        paymentMethod:
          type: string
        paystackReference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, deliveryAddress, items, paymentMethod]
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        deliveryAddress:
          type: string
          minLength: 1
        city:
          type: string
        state:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        paymentMethod:
          type: string
        notes:
          type: string
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        paymentStatus:
          type: string
    OrderListResponse:
      type: object
      required: [orders, total, page]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
...